Not right; now `onLoad` is called twice after loading from disk
